moved utility functions from name_generator to name_utils
