+++ name_generator.py
@@ -1,90 +1,14 @@
-# TODO: 
-# Write training loop
-# Write evaluation and testing scripts
-
-
-import glob
-import unicodedata
-import string
-import random
+import name_utils as utils
 from tqdm import *
 
 import torch
 from torch.autograd import Variable
 
 # Constants
-allowed_chars = string.ascii_letters + " .,;'"
-n_chars = len(allowed_chars)
 n_hidden = 128
 learning_rate = 0.0005
-training_iterations = 10000
+training_iterations = 5000
 print_every = 1000
-
-# Converts unicode strings into strings containing only ascii
-# Characters. Accented letters will have their accents
-# removed and other unicode characters will be removed
-def unicode_to_ascii(s):
-    return "".join(
-            c for c in unicodedata.normalize('NFD', s)
-            if unicodedata.category(c) != "Mn"
-            and c in allowed_chars
-        )
-
-# Returns a list of paths matching a given pattern
-def find_files(path):
-    return glob.glob(path)
-
-# Opens a file of the given name and
-# returns a list of the lines in ascii format
-def read_lines(filename):
-    with open(filename, encoding="UTF-8") as lines:
-        names = lines.read().strip().split('\n')
-        return [unicode_to_ascii(name) for name in names]
-
-# Declare data containers
-category_lines = {}
-categories = []
-
-# Load data into containers
-for filename in find_files("name_data/names/*.txt"):
-    category = filename.split('/')[-1].split('.')[0]
-    categories.append(category)
-    lines = read_lines(filename)
-    category_lines[category] = lines
-
-n_categories = len(categories)
-
-# Helper methods for converting strings to onehot tensors
-def letter_to_variable(letter):
-    one_hot_tensor = torch.zeros(1, n_chars)
-    index = allowed_chars.find(letter)
-    one_hot_tensor[0][index] = 1
-    return Variable(one_hot_tensor)
-
-def name_to_variable(name):
-    one_hot_tensor = torch.zeros(len(name), 1, n_chars)
-    # Loop over and add ones at a corresponding index
-    for i in range(len(name)):
-        index = allowed_chars.find(name[i])
-        one_hot_tensor[i][0][index] = 1
-    return Variable(one_hot_tensor)
-
-def name_to_target(name):
-    name = name + " "
-    indicies = [allowed_chars.find(char) for char in name]
-    target = Variable(torch.LongTensor(indicies))
-    return target
-
-def category_to_variable(category):
-    one_hot_tensor = torch.zeros(1, n_categories)
-    
-    if type(category) is int:
-        one_hot_tensor[0][category_index] = 1
-    elif type(category) is str:
-        index = categories.index(category)
-        one_hot_tensor[0][index] = 1
-    
-    return Variable(one_hot_tensor)
 
 
 # Define Generative Model
@@ -94,8 +18,8 @@ class Generator(torch.nn.Module):
         self.hidden_size = hidden_size
 
         # parameters
-        self.c2i = torch.nn.Linear(input_size + hidden_size + n_categories, output_size)
-        self.c2h = torch.nn.Linear(input_size + hidden_size + n_categories, hidden_size)
+        self.c2i = torch.nn.Linear(input_size + hidden_size + utils.n_categories, output_size)
+        self.c2h = torch.nn.Linear(input_size + hidden_size + utils.n_categories, hidden_size)
         self.i2o = torch.nn.Linear(output_size + hidden_size, output_size)
         # Read c2i as combined value to intermediate value;
         #      c2h as combined value to hidden value
@@ -124,31 +48,25 @@ class Generator(torch.nn.Module):
         return Variable(torch.zeros(1, self.hidden_size))
 
 # Training
-model = Generator(n_chars, n_hidden, n_chars)
+model = Generator(utils.n_chars, n_hidden, utils.n_chars)
 criterion = torch.nn.NLLLoss()
 optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
 
-def get_random_example():
-   rand_category_index = random.randint(0, n_categories - 1)
-   category = categories[rand_category_index]
-   rand_line_index = random.randint(0, len(category) - 1)
-   line = category_lines[category][rand_line_index]
-   return line, category, rand_category_index
 
 loss_accumulator = 0
 for iteration in tqdm(range(training_iterations)):
     # Get training example
-    line, category, category_index = get_random_example()
+    line, category, category_index = utils.get_random_example()
 
     # Convert to torch.autograd variables
-    name_variable = name_to_variable(line)
-    category_variable = category_to_variable(category_index)
-    target = name_to_target(line)
+    name_variable = utils.name_to_variable(line)
+    category_variable = utils.category_to_variable(category_index)
+    target = utils.name_to_target(line)
 
     # Append special start char to name
     name_variable = torch.cat(
             (
-                torch.unsqueeze(letter_to_variable(";"),0),
+                torch.unsqueeze(utils.letter_to_variable(";"),0),
                 name_variable
             ),0)
 
@@ -179,13 +97,13 @@ for iteration in tqdm(range(training_iterations)):
 # Test Generation
 def generate(category):
     # Create start token
-    prev_char = letter_to_variable(";")
+    prev_char = utils.letter_to_variable(";")
 
     # Initialize a new hidden state
     hidden_state = model.init_hidden()
 
     # Transform inputs into torch autograd variables
-    category = category_to_variable(category)
+    category = utils.category_to_variable(category)
 
     # Iterate over results from model
     result = ""
@@ -194,8 +112,8 @@ def generate(category):
         out, hidden_state = model(category, prev_char, hidden_state)
         
         char_index = torch.max(out, 1)[1].data[0][0]
-        char = allowed_chars[char_index]
-        prev_char = letter_to_variable(char)
+        char = utils.allowed_chars[char_index]
+        prev_char = utils.letter_to_variable(char)
 
         if char is " " or counter > 20:
             break
